+++ cmd/fofa/cmd/search.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"fmt"
 	"github.com/lubyruffy/gofofa/pkg/outformats"
+	"github.com/sirupsen/logrus"
 	"github.com/urfave/cli/v2"
 	"io"
 	"os"
@@ -60,6 +61,15 @@ var searchCmd = &cli.Command{
 	Action: SearchAction,
 }
 
+func hasBodyField(fields []string) bool {
+	for _, f := range fields {
+		if f == "body" {
+			return true
+		}
+	}
+	return false
+}
+
 // SearchAction search action
 func SearchAction(ctx *cli.Context) error {
 	// valid same config
@@ -88,15 +98,20 @@ func SearchAction(ctx *cli.Context) error {
 
 	// gen writer
 	var writer outformats.OutWriter
-	switch format {
-	case "csv":
-		writer = outformats.NewCSVWriter(outTo)
-	case "json":
+	if hasBodyField(fields) && format == "csv" {
+		logrus.Warnln("fields contains body, so change format to json")
 		writer = outformats.NewJSONWriter(outTo, fields)
-	case "xml":
-		writer = outformats.NewXMLWriter(outTo, fields)
-	default:
-		return fmt.Errorf("unknown format: %s", format)
+	} else {
+		switch format {
+		case "csv":
+			writer = outformats.NewCSVWriter(outTo)
+		case "json":
+			writer = outformats.NewJSONWriter(outTo, fields)
+		case "xml":
+			writer = outformats.NewXMLWriter(outTo, fields)
+		default:
+			return fmt.Errorf("unknown format: %s", format)
+		}
 	}
 
 	// do search